path_selection service
